Hr6
Anagram program:
Two strings, a and b , are called anagrams if they contain all the same characters in the same frequencies. For example, the anagrams of CAT are CAT, ACT, TAC, TCA, ATC, and CTA.

Complete the function in the editor. If  a and b  are case-insensitive anagrams, print "Anagrams"; otherwise, print "Not Anagrams" instead.